fix to show toc when directly loading tokens page
