+++ apps/dialtone-documentation/docs/.vuepress/theme/components/Page.vue
@@ -1,7 +1,7 @@
 <template>
   <div
     class="d-d-grid d-jc-center"
-    :class="[gridClass, tokensPageClass]"
+    :class="gridClass"
   >
     <div class="d-p24 lg:d-pr24 lg:d-pt64">
       <page-header />
@@ -55,7 +55,7 @@
       </footer>
     </div>
     <div class="d-ps-relative d-ga-toc">
-      <page-toc v-if="!isMobile && includeToc" />
+      <page-toc v-if="!isMobile && includeToc" :headers="headers" />
     </div>
   </div>
 </template>
@@ -63,7 +63,7 @@
 <script setup>
 import PageHeader from '../components/PageHeader.vue';
 import PageToc from '../components/PageToc.vue';
-import { computed } from 'vue';
+import { computed, watch, ref, nextTick } from 'vue';
 import { useRoute } from 'vue-router';
 import { usePageData } from '@vuepress/client';
 import { useThemeLocaleData } from '@vuepress/plugin-theme-data/client';
@@ -84,6 +84,9 @@ const props = defineProps({
     required: true,
   },
 });
+
+const headers = ref(null);
+
 const lastUpdated = computed(() => {
   const date = new Date(usePageData().value.git.updatedTime);
   return new Intl.DateTimeFormat('en-US', { dateStyle: 'full' }).format(date);
@@ -99,12 +102,13 @@ const includeToc = computed(() => {
   // get the item that matches the current route from site-nav without cosidering the last '/'
   const key = Object.keys(items).filter(item => route.path.includes(item.replace(/\/$/, '')));
   if (!items[key] || !Array.isArray(items[key])) return false;
-  const headers = usePageData().value.headers;
-  return headers?.length > 0 || localStorage.getItem('filteredHeaders');
+  return headers.value && headers.value.length > 0;
 });
 
-const tokensPageClass = computed(() => {
-  if (route.path.includes('tokens')) return 'tokens-page';
-  return '';
-});
+watch(route, async () => {
+  await nextTick();
+  const filteredHeaders = JSON.parse(localStorage.getItem('filteredHeaders'));
+  const pageHeaders = usePageData().value.headers;
+  headers.value = filteredHeaders ?? pageHeaders;
+}, { flush: 'pre', immediate: true, deep: true });
 </script>
