+++ apps/dialtone-documentation/docs/.vuepress/theme/components/PageToc.vue
@@ -57,13 +57,17 @@
 </template>
 
 <script setup>
-import { ref, watch } from 'vue';
-import { usePageData } from '@vuepress/client';
 import { useRoute } from 'vue-router';
 import TocItem from './TocItem.vue';
 
+defineProps({
+  headers: {
+    type: Array,
+    default: null,
+  },
+});
+
 const route = useRoute();
-const headers = ref(null);
 
 function isHeaderActive (header) {
   const links = [header.link, ...header.children.map(child => child.link)];
@@ -73,12 +77,6 @@ function isHeaderActive (header) {
 function isItemActive (item) {
   return item.link === route.hash;
 }
-
-watch(route, () => {
-  const filteredHeaders = JSON.parse(localStorage.getItem('filteredHeaders'));
-  const pageHeaders = usePageData().value.headers;
-  headers.value = filteredHeaders ?? pageHeaders;
-}, { flush: 'pre', immediate: true, deep: true });
 </script>
 
 <style lang="less" scoped>
